+++ contracts/staking/linear/LinearPool.sol
@@ -4,6 +4,7 @@ pragma solidity 0.8.28;
 import {ContractOwnership} from "./../../access/ContractOwnership.sol";
 import {AccessControl} from "./../../access/AccessControl.sol";
 import {ReentrancyGuard} from "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
+import {TokenRecovery} from "./../../security/TokenRecovery.sol";
 import {Context} from "@openzeppelin/contracts/utils/Context.sol";
 import {ForwarderRegistryContextBase} from "./../../metatx/base/ForwarderRegistryContextBase.sol";
 import {ForwarderRegistryContext} from "./../../metatx/ForwarderRegistryContext.sol";
@@ -14,7 +15,14 @@ import {IForwarderRegistry} from "./../../metatx/interfaces/IForwarderRegistry.s
 
 // design inspired from https://github.com/k06a/Unipool/blob/master/contracts/Unipool.sol
 
-abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, ForwarderRegistryContext {
+/// @title Linear rewards distribution staking pool.
+/// @notice Implements the base logic for linear reward pools, while the nature of the staking and rewards is left to the deriving contracts.
+/// @notice Stakes, whether fungible or non-fungible, map to an amount of "stake points", then used to compute the user rewards share.
+/// @notice NB: Reentrancy guards are used to protect the stake and withdraw functions, the implementation being unknown.
+/// @notice If the deriving contract's implementation does not present reentrancy elements, the guards can be dropped.
+/// @notice NB: This contract inherits TokenRecovery functions. In the likely event that the deriving contract does keep tokens in stake,
+/// @notice the corresponding functions must be overriden to prevent recovering tokens legitimately staked in the contract.
+abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, TokenRecovery, ForwarderRegistryContext {
     using AccessControlStorage for AccessControlStorage.Layout;
     using SafeERC20 for IERC20;
 
@@ -57,11 +65,15 @@ abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, For
         }
     }
 
+    /// @notice Returns the last time rewards are applicable.
+    /// @return The minimum of the current block timestamp and the distribution end.
     function lastTimeRewardApplicable() public view returns (uint256) {
         uint256 currentDistributionEnd = distributionEnd;
         return block.timestamp < currentDistributionEnd ? block.timestamp : currentDistributionEnd;
     }
 
+    /// @notice Returns the current reward per stake point.
+    /// @return The sum of the last stored value and the new rewards since the last update
     function rewardPerStakePoint() public view returns (uint256) {
         uint256 currentTotalStaked = totalStaked;
         if (currentTotalStaked == 0) {
@@ -70,14 +82,30 @@ abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, For
         return rewardPerStakePointStored + (((lastTimeRewardApplicable() - lastUpdated) * rewardRate * SCALING_FACTOR) / currentTotalStaked);
     }
 
+    /// @notice Returns the amount of rewards earned by the account.
+    /// @param account The address of the account to check.
+    /// @return The account's stake points times the difference between the current reward per stake point and the last paid reward per stake point.
     function earned(address account) public view returns (uint256) {
         return (staked[account] * (rewardPerStakePoint() - rewardPerStakePointPaid[account])) / SCALING_FACTOR + rewards[account];
     }
 
+    /// @notice Stakes to the pool.
+    /// @param stakeData The data to be used for the stake (encoding freely determined by the deriving contracts).
+    /// @dev Reverts with {ReentrancyGuardReentrantCall} if the function is re-entered.
+    /// @dev Reverts with {InvalidStakeAmount} if the stake amount is 0.
+    /// @dev Emits a {Staked} event with the staker address, stakeData and stake points.
+    /// @dev The stakeData is passed to the _computeStake function, which must be implemented in the deriving contract.
+    /// @dev The stakeData is not used in the base implementation, but it is passed to the event for convenience.
     function stake(bytes calldata stakeData) public payable virtual nonReentrant {
         _stake(_msgSender(), stakeData);
     }
 
+    /// @notice Stakes to the pool.
+    /// @param staker The address of the staker.
+    /// @param stakeData The data to be used for the stake (encoding freely determined by the deriving contracts).
+    /// @dev Reverts with {InvalidStakeAmount} if the stake amount is 0.
+    /// @dev Emits a {Staked} event with the staker address, stakeData and stake points.
+    /// @dev The stakeData is passed to the _computeStake function, which must be implemented in the deriving contract.
     function _stake(address staker, bytes memory stakeData) internal virtual {
         _updateReward(staker);
         uint256 stakePoints = _computeStake(staker, stakeData);
@@ -87,10 +115,25 @@ abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, For
         emit Staked(staker, stakeData, stakePoints);
     }
 
+    /// @notice Withdraws from the pool.
+    /// @param withdrawData The data to be used for the withdraw (encoding freely determined by the deriving contracts).
+    /// @dev Reverts with {ReentrancyGuardReentrantCall} if the function is re-entered.
+    /// @dev Reverts with {InvalidWithdrawAmount} if the withdraw amount is 0.
+    /// @dev Reverts with {NotEnoughStake} if the staker does not have enough stake points to withdraw.
+    /// @dev Emits a {Withdrawn} event with the staker address, withdrawData and stake points.
+    /// @dev The withdrawData is passed to the _computeWithdraw function, which must be implemented in the deriving contract.
+    /// @dev The withdrawData is not used in the base implementation, but it is passed to the event for convenience.
     function withdraw(bytes calldata withdrawData) public virtual nonReentrant {
         _withdraw(_msgSender(), withdrawData);
     }
 
+    /// @notice Withdraws from the pool.
+    /// @param staker The address of the staker.
+    /// @param withdrawData The data to be used for the withdraw (encoding freely determined by the deriving contracts).
+    /// @dev Reverts with {InvalidWithdrawAmount} if the withdraw amount is 0.
+    /// @dev Reverts with {NotEnoughStake} if the staker does not have enough stake points to withdraw.
+    /// @dev Emits a {Withdrawn} event with the staker address, withdrawData and stake points.
+    /// @dev The withdrawData is passed to the _computeWithdraw function, which must be implemented in the deriving contract.
     function _withdraw(address staker, bytes memory withdrawData) internal virtual {
         _updateReward(staker);
         uint256 stakePoints = _computeWithdraw(staker, withdrawData);
@@ -105,6 +148,9 @@ abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, For
         emit Withdrawn(staker, withdrawData, stakePoints);
     }
 
+    /// @notice Claims the rewards for the sender.
+    /// @dev Emits a {Claimed} event with the staker address, claimData and reward.
+    /// @dev The claimData is generated by the _computeClaim function, which must be implemented in the deriving contract.
     function claim() public virtual {
         address staker = _msgSender();
         _updateReward(staker);
@@ -116,6 +162,17 @@ abstract contract LinearPool is ILinearPool, AccessControl, ReentrancyGuard, For
         }
     }
 
+    /// @notice Adds rewards to the pool.
+    /// @notice If there is an ongoing distribution, the new rewards are added to the current distribution:
+    /// @notice - If the new distribution ends before the current one, the new rewards are added to the current distribution.
+    /// @notice - If the new distribution ends after the current one, the remaining rewards are added to the new distribution.
+    /// @param reward The amount of rewards to be added.
+    /// @param duration The duration of the rewards distribution.
+    /// @dev Reverts with {NotRoleHolder} if the sender does not have the REWARDER_ROLE.
+    /// @dev Reverts with {InvalidRewardAmount} if the reward amount is 0.
+    /// @dev Reverts with {InvalidDuration} if the duration is 0.
+    /// @dev Reverts with {RewardTooSmallForDuration} if the reward is too small for the duration.
+    /// @dev Emits a {RewardAdded} event with the rewarder address, reward amount, duration and dust.
     function addReward(uint256 reward, uint256 duration) public payable virtual {
         address rewarder = _msgSender();
         AccessControlStorage.layout().enforceHasRole(REWARDER_ROLE, rewarder);